requirements.txt
Flask==0.12.2
flask-jsonpify
flask-sqlalchemy
flask-restful
lightfm
eli5
pandas
psycopg2-binary
xgboost
